pyproject.toml
[project]
name = "asyncpytemplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncpg>=0.30.0",
    "dishka>=1.7.2",
    "dotenv>=0.9.9",
    "fastapi>=0.121.3",
    "fastapi-msgspec>=0.1.0",
    "psycopg2>=2.9.11",
    "pytz>=2025.2",
    "redis>=7.1.0",
    "yoyo-migrations>=9.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]

